[intel-npu] Unify options under 1 header/source in preparation for refactoring. Updating all dependencies and occurencies. (separation was just adding useless complexity and ambiguity)
