+++ src/plugins/intel_npu/src/al/include/intel_npu/al/config/npuw.hpp
@@ -6,9 +6,9 @@
 
 #include <thread>
 
-#include "common.hpp"
 #include "npu_private_properties.hpp"
 #include "npuw_private_properties.hpp"
+#include "options.hpp"
 
 namespace intel_npu {
 
@@ -18,13 +18,19 @@ namespace intel_npu {
 
 void registerNPUWOptions(OptionsDesc& desc);
 
-#define DEFINE_OPT(Name, Type, DefaultValue, PropertyKey, Mode)                     \
-    struct Name final : OptionBase<Name, Type> {                                    \
-        static std::string_view key() { return ov::intel_npu::PropertyKey.name(); } \
-                                                                                    \
-        static Type defaultValue() { return DefaultValue; }                         \
-                                                                                    \
-        static OptionMode mode() { return OptionMode::Mode; }                       \
+#define DEFINE_OPT(Name, Type, DefaultValue, PropertyKey, Mode) \
+    struct Name final : OptionBase<Name, Type> {                \
+        static std::string_view key() {                         \
+            return ov::intel_npu::PropertyKey.name();           \
+        }                                                       \
+                                                                \
+        static Type defaultValue() {                            \
+            return DefaultValue;                                \
+        }                                                       \
+                                                                \
+        static OptionMode mode() {                              \
+            return OptionMode::Mode;                            \
+        }                                                       \
     };
 
 DEFINE_OPT(NPU_USE_NPUW, bool, false, use_npuw, CompileTime);
