+++ src/plugins/intel_npu/src/backend/src/zero_infer_request.cpp
@@ -4,9 +4,7 @@
 
 #include "zero_infer_request.hpp"
 
-#include "intel_npu/al/config/common.hpp"
-#include "intel_npu/al/config/compiler.hpp"
-#include "intel_npu/al/config/runtime.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "intel_npu/al/itt.hpp"
 #include "intel_npu/al/prefix.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
