+++ src/plugins/intel_npu/src/backend/src/zero_wrappers.cpp
@@ -4,7 +4,7 @@
 
 #include "zero_wrappers.hpp"
 
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "intel_npu/utils/zero/zero_api.hpp"
 #include "zero_types.hpp"
