+++ src/plugins/intel_npu/src/plugin/src/backends.cpp
@@ -8,7 +8,7 @@
 #include <memory>
 
 #include "device_helpers.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 #if defined(ENABLE_ZEROAPI_BACKEND)
 #    include "zero_backend.hpp"
