+++ src/plugins/intel_npu/src/plugin/src/compiler.cpp
@@ -4,7 +4,7 @@
 
 #include "compiler.hpp"
 
-#include "intel_npu/al/config/compiler.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "intel_npu/al/itt.hpp"
 #include "intel_npu/utils/logger/logger.hpp"
 #include "npu_private_properties.hpp"
