+++ src/plugins/intel_npu/src/plugin/src/plugin.cpp
@@ -9,10 +9,8 @@
 #include "compiled_model.hpp"
 #include "compiler.hpp"
 #include "device_helpers.hpp"
-#include "intel_npu/al/config/common.hpp"
-#include "intel_npu/al/config/compiler.hpp"
 #include "intel_npu/al/config/npuw.hpp"
-#include "intel_npu/al/config/runtime.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "intel_npu/al/itt.hpp"
 #include "npuw/compiled_model.hpp"
 #include "openvino/op/constant.hpp"
