+++ src/plugins/intel_npu/tests/functional/behavior/fail_gracefully_forward_compatibility.cpp
@@ -3,9 +3,10 @@
 //
 
 #include "behavior/fail_gracefully_forward_compatibility.hpp"
-#include "common/utils.hpp"
+
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "common/utils.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 using namespace ov::test::behavior;
 
@@ -17,7 +18,8 @@ namespace {
 
 const std::vector<ov::AnyMap> configs = {{}};
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTest, FailGracefullyTest,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTest,
+                         FailGracefullyTest,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          FailGracefullyTest::getTestCaseName);
