+++ src/plugins/intel_npu/tests/functional/behavior/npu_driver_compiler_adapter/inputs_outputs.cpp
@@ -3,17 +3,19 @@
 //
 
 #include "inputs_outputs.hpp"
-#include "common/utils.hpp"
+
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "common/utils.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 using namespace ov::test::behavior;
 
 const std::vector<ov::AnyMap> configs = {
-        {{ov::intel_npu::compiler_type(ov::intel_npu::CompilerType::DRIVER)}},
+    {{ov::intel_npu::compiler_type(ov::intel_npu::CompilerType::DRIVER)}},
 };
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTest, DriverCompilerAdapterInputsOutputsTestNPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTest,
+                         DriverCompilerAdapterInputsOutputsTestNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          DriverCompilerAdapterInputsOutputsTestNPU::getTestCaseName);
