+++ src/plugins/intel_npu/tests/functional/behavior/npu_driver_compiler_adapter/properties_compatibility.cpp
@@ -4,7 +4,7 @@
 
 #include "base/ov_behavior_test_utils.hpp"
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "shared_test_classes/subgraph/split_conv_concat.hpp"
 
 using CompilationParams = std::tuple<std::string,  // Device name
