+++ src/plugins/intel_npu/tests/functional/behavior/remote_tensor_tests/dx12_remote_run.cpp
@@ -6,7 +6,7 @@
 
 #include "common/npu_test_env_cfg.hpp"
 #include "common/utils.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 #ifdef _WIN32
 #    ifdef ENABLE_DX12
