+++ src/plugins/intel_npu/tests/functional/behavior/work_with_devices.cpp
@@ -3,16 +3,18 @@
 //
 
 #include "behavior/work_with_devices.hpp"
+
 #include "common/utils.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 namespace {
 
 const std::vector<ov::AnyMap> configs = {
-        {{ov::log::level(ov::log::Level::DEBUG)}},
+    {{ov::log::level(ov::log::Level::DEBUG)}},
 };
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTest, TestCompiledModelNPU,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTest,
+                         TestCompiledModelNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          TestCompiledModelNPU::getTestCaseName);
