+++ src/plugins/intel_npu/tests/functional/behavior/work_with_devices.hpp
@@ -6,15 +6,14 @@
 
 #include <base/ov_behavior_test_utils.hpp>
 #include <cstring>
-#include <intel_npu/al/config/compiler.hpp>
+#include <intel_npu/al/config/options.hpp>
 #include <string>
 #include <vector>
 
 #include "common/functions.h"
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "npu_private_properties.hpp"
-
 
 using CompilerType = ov::intel_npu::CompilerType;
