+++ src/plugins/intel_npu/tests/functional/common/npu_test_env_cfg.cpp
@@ -7,9 +7,8 @@
 #include <cstdlib>
 #include <stdexcept>
 
-#include "intel_npu/al/config/common.hpp"
 #include "intel_npu/al/config/config.hpp"
-
+#include "intel_npu/al/config/options.hpp"
 
 namespace ov::test::utils {
