+++ src/plugins/intel_npu/tests/functional/internal/overload/compile_and_infer.hpp
@@ -9,7 +9,7 @@
 #include <sstream>
 
 #include "base/ov_behavior_test_utils.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "npu_private_properties.hpp"
 #include "openvino/core/except.hpp"
 #include "openvino/opsets/opset8.hpp"
