+++ src/plugins/intel_npu/tests/functional/internal/overload/compiled_model/property.cpp
@@ -2,40 +2,38 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "behavior/compiled_model/properties.hpp"
-#include "common/utils.hpp"
-#include "common/npu_test_env_cfg.hpp"
-#include "common_test_utils/subgraph_builders/conv_pool_relu.hpp"
-#include "intel_npu/al/config/common.hpp"
-
 #include <openvino/runtime/intel_npu/properties.hpp>
-
 #include <vector>
+
+#include "behavior/compiled_model/properties.hpp"
+#include "common/npu_test_env_cfg.hpp"
+#include "common/utils.hpp"
+#include "common_test_utils/subgraph_builders/conv_pool_relu.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 using namespace ov::test::behavior;
 
 namespace {
 
 std::vector<std::pair<std::string, ov::Any>> exe_network_supported_properties = {
-        {ov::hint::num_requests.name(), ov::Any(8)},
-        {ov::hint::enable_cpu_pinning.name(), ov::Any(true)},
-        {ov::hint::performance_mode.name(), ov::Any(ov::hint::PerformanceMode::THROUGHPUT)},
-        {ov::enable_profiling.name(), ov::Any(true)},
-        {ov::device::id.name(), ov::Any(ov::test::utils::getDeviceNameID(ov::test::utils::getDeviceName()))},
-        {ov::optimal_number_of_infer_requests.name(), ov::Any(2)},
+    {ov::hint::num_requests.name(), ov::Any(8)},
+    {ov::hint::enable_cpu_pinning.name(), ov::Any(true)},
+    {ov::hint::performance_mode.name(), ov::Any(ov::hint::PerformanceMode::THROUGHPUT)},
+    {ov::enable_profiling.name(), ov::Any(true)},
+    {ov::device::id.name(), ov::Any(ov::test::utils::getDeviceNameID(ov::test::utils::getDeviceName()))},
+    {ov::optimal_number_of_infer_requests.name(), ov::Any(2)},
 };
 
 std::vector<std::pair<std::string, ov::Any>> exe_network_immutable_properties = {
-        {std::make_pair(ov::optimal_number_of_infer_requests.name(), ov::Any(2))},
-        {std::make_pair(ov::hint::enable_cpu_pinning.name(), ov::Any(false))},
-        {std::make_pair(ov::supported_properties.name(), ov::Any("deadbeef"))},
-        {std::make_pair(ov::model_name.name(), ov::Any("deadbeef"))}};
+    {std::make_pair(ov::optimal_number_of_infer_requests.name(), ov::Any(2))},
+    {std::make_pair(ov::hint::enable_cpu_pinning.name(), ov::Any(false))},
+    {std::make_pair(ov::supported_properties.name(), ov::Any("deadbeef"))},
+    {std::make_pair(ov::model_name.name(), ov::Any("deadbeef"))}};
 
 // ExecutableNetwork Properties tests
-class ClassExecutableNetworkGetPropertiesTestNPU :
-        public OVCompiledModelPropertiesBase,
-        public ::testing::WithParamInterface<
-                std::tuple<std::string, std::pair<std::string, ov::Any>>> {
+class ClassExecutableNetworkGetPropertiesTestNPU
+    : public OVCompiledModelPropertiesBase,
+      public ::testing::WithParamInterface<std::tuple<std::string, std::pair<std::string, ov::Any>>> {
 protected:
     std::string deviceName;
     std::string configKey;
@@ -52,9 +50,7 @@ public:
         model = ov::test::utils::make_conv_pool_relu();
     }
     static std::string getTestCaseName(
-            testing::TestParamInfo<
-                    std::tuple<std::string, std::pair<std::string, ov::Any>>>
-                    obj) {
+        testing::TestParamInfo<std::tuple<std::string, std::pair<std::string, ov::Any>>> obj) {
         std::string targetDevice;
         std::pair<std::string, ov::Any> configuration;
         std::tie(targetDevice, configuration) = obj.param;
@@ -73,8 +69,7 @@ using ClassExecutableNetworkTestSuite1NPU = ClassExecutableNetworkGetPropertiesT
 TEST_P(ClassExecutableNetworkTestSuite1NPU, PropertyIsSupportedAndImmutableAndGet) {
     std::vector<ov::PropertyName> properties;
 
-    ov::CompiledModel exeNetwork =
-            ie.compile_model(model, deviceName);
+    ov::CompiledModel exeNetwork = ie.compile_model(model, deviceName);
     OV_ASSERT_NO_THROW(properties = exeNetwork.get_property(ov::supported_properties));
 
     auto it = find(properties.cbegin(), properties.cend(), configKey);
@@ -95,8 +90,7 @@ using ClassExecutableNetworkTestSuite2NPU = ClassExecutableNetworkGetPropertiesT
 TEST_P(ClassExecutableNetworkTestSuite2NPU, PropertyIsSupportedAndImmutableAndCanNotSet) {
     std::vector<ov::PropertyName> properties;
 
-    ov::CompiledModel exeNetwork =
-            ie.compile_model(model, deviceName);
+    ov::CompiledModel exeNetwork = ie.compile_model(model, deviceName);
     OV_ASSERT_NO_THROW(properties = exeNetwork.get_property(ov::supported_properties));
 
     auto it = find(properties.cbegin(), properties.cend(), configKey);
@@ -106,7 +100,8 @@ TEST_P(ClassExecutableNetworkTestSuite2NPU, PropertyIsSupportedAndImmutableAndCa
     ASSERT_THROW(exeNetwork.set_property({{configKey, configValue}}), ov::Exception);
 }
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassExecutableNetworkTestSuite2NPU, ClassExecutableNetworkTestSuite2NPU,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassExecutableNetworkTestSuite2NPU,
+                         ClassExecutableNetworkTestSuite2NPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
                                             ::testing::ValuesIn(exe_network_immutable_properties)),
                          ClassExecutableNetworkTestSuite2NPU::getTestCaseName);
@@ -116,46 +111,46 @@ INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassExecutableNetworkTestSuite2NPU
 namespace {
 
 std::vector<std::pair<std::string, ov::Any>> plugin_public_mutable_properties = {
-        {ov::hint::num_requests.name(), ov::Any(5)},
-        {ov::enable_profiling.name(), ov::Any(true)},
-        {ov::compilation_num_threads.name(), ov::Any(1)},
-        {ov::hint::performance_mode.name(), ov::Any(ov::hint::PerformanceMode::THROUGHPUT)},
-        {ov::hint::enable_cpu_pinning.name(), ov::Any(true)},
-        {ov::log::level.name(), ov::Any(ov::log::Level::DEBUG)},
-        {ov::device::id.name(), ov::Any(ov::test::utils::getDeviceNameID(ov::test::utils::getDeviceName()))},
+    {ov::hint::num_requests.name(), ov::Any(5)},
+    {ov::enable_profiling.name(), ov::Any(true)},
+    {ov::compilation_num_threads.name(), ov::Any(1)},
+    {ov::hint::performance_mode.name(), ov::Any(ov::hint::PerformanceMode::THROUGHPUT)},
+    {ov::hint::enable_cpu_pinning.name(), ov::Any(true)},
+    {ov::log::level.name(), ov::Any(ov::log::Level::DEBUG)},
+    {ov::device::id.name(), ov::Any(ov::test::utils::getDeviceNameID(ov::test::utils::getDeviceName()))},
 };
 
 std::vector<std::pair<std::string, ov::Any>> plugin_internal_mutable_properties = {
-        {ov::intel_npu::compilation_mode_params.name(), ov::Any("use-user-precision=false propagate-quant-dequant=0")},
-        {ov::intel_npu::dma_engines.name(), ov::Any(1)},
-        {ov::intel_npu::compiler_type.name(), ov::Any(ov::intel_npu::CompilerType::MLIR)},
-        {ov::intel_npu::platform.name(), ov::Any(ov::intel_npu::Platform::AUTO_DETECT)},
-        {ov::intel_npu::compilation_mode.name(), ov::Any("DefaultHW")},
-        {ov::intel_npu::max_tiles.name(), ov::Any(8)},
-        {ov::intel_npu::stepping.name(), ov::Any(4)},
-        {ov::intel_npu::dpu_groups.name(), ov::Any(2)},
-        {ov::intel_npu::use_elf_compiler_backend.name(), ov::Any(ov::intel_npu::ElfCompilerBackend::YES)},
-        {ov::intel_npu::create_executor.name(), ov::Any(0)},
+    {ov::intel_npu::compilation_mode_params.name(), ov::Any("use-user-precision=false propagate-quant-dequant=0")},
+    {ov::intel_npu::dma_engines.name(), ov::Any(1)},
+    {ov::intel_npu::compiler_type.name(), ov::Any(ov::intel_npu::CompilerType::MLIR)},
+    {ov::intel_npu::platform.name(), ov::Any(ov::intel_npu::Platform::AUTO_DETECT)},
+    {ov::intel_npu::compilation_mode.name(), ov::Any("DefaultHW")},
+    {ov::intel_npu::max_tiles.name(), ov::Any(8)},
+    {ov::intel_npu::stepping.name(), ov::Any(4)},
+    {ov::intel_npu::dpu_groups.name(), ov::Any(2)},
+    {ov::intel_npu::use_elf_compiler_backend.name(), ov::Any(ov::intel_npu::ElfCompilerBackend::YES)},
+    {ov::intel_npu::create_executor.name(), ov::Any(0)},
 };
 
 std::vector<std::pair<std::string, ov::Any>> plugin_public_immutable_properties = {
-        {ov::device::uuid.name(), ov::Any("deadbeef")},
-        {ov::supported_properties.name(), {ov::device::full_name.name()}},
-        {ov::num_streams.name(), ov::Any(ov::streams::Num(4))},
-        {ov::available_devices.name(), ov::Any(std::vector<std::string>{"deadbeef"})},
-        {ov::device::capabilities.name(), ov::Any(std::vector<std::string>{"deadbeef"})},
-        {ov::range_for_async_infer_requests.name(),
-         ov::Any(std::tuple<unsigned int, unsigned int, unsigned int>{0, 10, 1})},
-        {ov::range_for_streams.name(), ov::Any(std::tuple<unsigned int, unsigned int>{0, 10})},
-        {ov::optimal_number_of_infer_requests.name(), ov::Any(4)},
-        {ov::intel_npu::device_alloc_mem_size.name(), ov::Any(2)},
-        {ov::intel_npu::device_total_mem_size.name(), ov::Any(2)},
+    {ov::device::uuid.name(), ov::Any("deadbeef")},
+    {ov::supported_properties.name(), {ov::device::full_name.name()}},
+    {ov::num_streams.name(), ov::Any(ov::streams::Num(4))},
+    {ov::available_devices.name(), ov::Any(std::vector<std::string>{"deadbeef"})},
+    {ov::device::capabilities.name(), ov::Any(std::vector<std::string>{"deadbeef"})},
+    {ov::range_for_async_infer_requests.name(),
+     ov::Any(std::tuple<unsigned int, unsigned int, unsigned int>{0, 10, 1})},
+    {ov::range_for_streams.name(), ov::Any(std::tuple<unsigned int, unsigned int>{0, 10})},
+    {ov::optimal_number_of_infer_requests.name(), ov::Any(4)},
+    {ov::intel_npu::device_alloc_mem_size.name(), ov::Any(2)},
+    {ov::intel_npu::device_total_mem_size.name(), ov::Any(2)},
 };
 
 // Plugin Properties tests
-class ClassPluginPropertiesTestNPU :
-        public OVCompiledModelPropertiesBase,
-        public ::testing::WithParamInterface<std::tuple<std::string, std::pair<std::string, ov::Any>>> {
+class ClassPluginPropertiesTestNPU
+    : public OVCompiledModelPropertiesBase,
+      public ::testing::WithParamInterface<std::tuple<std::string, std::pair<std::string, ov::Any>>> {
 protected:
     std::string deviceName;
     std::string configKey;
@@ -170,7 +165,7 @@ public:
         std::tie(configKey, configValue) = std::get<1>(GetParam());
     }
     static std::string getTestCaseName(
-            testing::TestParamInfo<std::tuple<std::string, std::pair<std::string, ov::Any>>> obj) {
+        testing::TestParamInfo<std::tuple<std::string, std::pair<std::string, ov::Any>>> obj) {
         std::string targetDevice;
         std::pair<std::string, ov::Any> configuration;
         std::tie(targetDevice, configuration) = obj.param;
@@ -202,7 +197,8 @@ TEST_P(ClassPluginPropertiesTestSuite0NPU, CanSetGetPublicMutableProperty) {
     ASSERT_EQ(retrieved_value.as<std::string>(), configValue.as<std::string>());
 }
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_ClassPluginPropertiesTestNPU, ClassPluginPropertiesTestSuite0NPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_ClassPluginPropertiesTestNPU,
+                         ClassPluginPropertiesTestSuite0NPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
                                             ::testing::ValuesIn(plugin_public_mutable_properties)),
                          ClassPluginPropertiesTestNPU::getTestCaseName);
@@ -218,7 +214,8 @@ TEST_P(ClassPluginPropertiesTestSuite1NPU, CanSetGetInternalMutableProperty) {
     ASSERT_EQ(retrieved_value.as<std::string>(), configValue.as<std::string>());
 }
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_ClassPluginPropertiesTestNPU, ClassPluginPropertiesTestSuite1NPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_ClassPluginPropertiesTestNPU,
+                         ClassPluginPropertiesTestSuite1NPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
                                             ::testing::ValuesIn(plugin_internal_mutable_properties)),
                          ClassPluginPropertiesTestNPU::getTestCaseName);
@@ -245,7 +242,8 @@ TEST_P(ClassPluginPropertiesTestSuite2NPU, CanNotSetImmutableProperty) {
     ASSERT_EQ(orig_value.as<std::string>(), after_value.as<std::string>());
 }
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassPluginPropertiesTest, ClassPluginPropertiesTestSuite2NPU,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassPluginPropertiesTest,
+                         ClassPluginPropertiesTestSuite2NPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
                                             ::testing::ValuesIn(plugin_public_immutable_properties)),
                          ClassPluginPropertiesTestNPU::getTestCaseName);
@@ -265,8 +263,9 @@ TEST_P(ClassPluginPropertiesTestSuite3NPU, CanGetPropertyWithOptionsNotAffecting
 
     ov::Any retrieved_value_with_options;
     OV_ASSERT_NO_THROW(retrieved_value_with_options = ie.get_property(
-                               deviceName, configKey,
-                               {{ov::hint::performance_mode.name(), ov::Any(ov::hint::PerformanceMode::THROUGHPUT)}}));
+                           deviceName,
+                           configKey,
+                           {{ov::hint::performance_mode.name(), ov::Any(ov::hint::PerformanceMode::THROUGHPUT)}}));
 
     ov::Any retrieved_value2;
     OV_ASSERT_NO_THROW(retrieved_value2 = ie.get_property(deviceName, configKey));
@@ -274,12 +273,14 @@ TEST_P(ClassPluginPropertiesTestSuite3NPU, CanGetPropertyWithOptionsNotAffecting
     ASSERT_EQ(retrieved_value.as<std::string>(), retrieved_value2.as<std::string>());
 }
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassPluginPropertiesOptsTest1NPU, ClassPluginPropertiesTestSuite3NPU,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassPluginPropertiesOptsTest1NPU,
+                         ClassPluginPropertiesTestSuite3NPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
                                             ::testing::ValuesIn(plugin_public_immutable_properties)),
                          ClassPluginPropertiesTestNPU::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_ClassPluginPropertiesOptsTest2NPU, ClassPluginPropertiesTestSuite3NPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_ClassPluginPropertiesOptsTest2NPU,
+                         ClassPluginPropertiesTestSuite3NPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
                                             ::testing::ValuesIn(plugin_public_mutable_properties)),
                          ClassPluginPropertiesTestNPU::getTestCaseName);
@@ -297,15 +298,12 @@ TEST_P(ClassPluginPropertiesTestSuite4NPU, CanNotSetGetInexistentProperty) {
 
     ASSERT_THROW(auto property1 = ie.get_property(deviceName, configKey), ov::Exception);
 
-    ASSERT_THROW(
-            ov::CompiledModel compiled_model1 = ie.compile_model(
-                    model, deviceName, {{configKey, configValue}}),
-            ov::Exception);
+    ASSERT_THROW(ov::CompiledModel compiled_model1 = ie.compile_model(model, deviceName, {{configKey, configValue}}),
+                 ov::Exception);
 
     ov::CompiledModel compiled_model2;
 
-    OV_ASSERT_NO_THROW(compiled_model2 =
-                               ie.compile_model(model, deviceName));
+    OV_ASSERT_NO_THROW(compiled_model2 = ie.compile_model(model, deviceName));
 
     ASSERT_THROW(compiled_model2.set_property({{configKey, configValue}}),
                  ov::Exception);  // Expect to throw due to unimplemented method
@@ -314,11 +312,12 @@ TEST_P(ClassPluginPropertiesTestSuite4NPU, CanNotSetGetInexistentProperty) {
                  ov::Exception);  // Expect to throw due to unsupported config
 }
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_ClassExecutableNetworkGetPropertiesTestNPU,
-                         ClassPluginPropertiesTestSuite4NPU,
-                         ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
-                                            ::testing::ValuesIn({std::make_pair<std::string, ov::Any>(
-                                                    "THISCONFIGKEYNOTEXIST", ov::Any("THISCONFIGVALUENOTEXIST"))})),
-                         ClassPluginPropertiesTestSuite4NPU::getTestCaseName);
+INSTANTIATE_TEST_SUITE_P(
+    smoke_BehaviorTests_ClassExecutableNetworkGetPropertiesTestNPU,
+    ClassPluginPropertiesTestSuite4NPU,
+    ::testing::Combine(::testing::Values(ov::test::utils::getDeviceName()),
+                       ::testing::ValuesIn({std::make_pair<std::string, ov::Any>("THISCONFIGKEYNOTEXIST",
+                                                                                 ov::Any("THISCONFIGVALUENOTEXIST"))})),
+    ClassPluginPropertiesTestSuite4NPU::getTestCaseName);
 
 }  // namespace
