+++ src/plugins/intel_npu/tests/functional/shared_tests_instances/behavior/ov_infer_request/inference_chaining.cpp
@@ -3,22 +3,25 @@
 //
 
 #include "behavior/ov_infer_request/inference_chaining.hpp"
-#include "common/utils.hpp"
+
 #include "common/npu_test_env_cfg.hpp"
+#include "common/utils.hpp"
 #include "common_test_utils/test_constants.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 using namespace ov::test::behavior;
 
 namespace {
 const std::vector<ov::AnyMap> configs = {{}};
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests, OVInferenceChaining,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests,
+                         OVInferenceChaining,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferenceChaining>);
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests, OVInferenceChainingStatic,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests,
+                         OVInferenceChainingStatic,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferenceChainingStatic>);
