+++ src/plugins/intel_npu/tests/functional/shared_tests_instances/behavior/ov_infer_request/io_tensor.cpp
@@ -2,12 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+#include "behavior/ov_infer_request/io_tensor.hpp"
+
 #include <vector>
 
-#include "behavior/ov_infer_request/io_tensor.hpp"
-#include "common/utils.hpp"
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "common/utils.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "overload/ov_infer_request/io_tensor.hpp"
 
 using namespace ov::test::behavior;
@@ -16,86 +17,120 @@ namespace {
 const std::vector<ov::AnyMap> configs = {{}};
 
 const std::vector<ov::AnyMap> multiConfigs = {
-        {ov::device::priorities(ov::test::utils::DEVICE_NPU),
-        ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
+    {ov::device::priorities(ov::test::utils::DEVICE_NPU), ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
 
 const std::vector<ov::AnyMap> autoConfigs = {
-        {ov::device::priorities(ov::test::utils::DEVICE_NPU),
-        ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
+    {ov::device::priorities(ov::test::utils::DEVICE_NPU), ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests, OVInferRequestIOTensorTest,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests,
+                         OVInferRequestIOTensorTest,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests, OVInferRequestIOTensorTest,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests,
+                         OVInferRequestIOTensorTest,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests, OVInferRequestIOTensorTestNPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests,
+                         OVInferRequestIOTensorTestNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests, OVInferRequestIOTensorTestNPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests,
+                         OVInferRequestIOTensorTestNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests, OVInferRequestIOTensorTest,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests,
+                         OVInferRequestIOTensorTest,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_AUTO),
                                             ::testing::ValuesIn(autoConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
 const std::vector<ov::element::Type> prcs = {
-        ov::element::boolean, ov::element::bf16, ov::element::f16, ov::element::f32, ov::element::f64, ov::element::i4,
-        ov::element::i8,      ov::element::i16,  ov::element::i32, ov::element::i64, ov::element::u1,  ov::element::u4,
-        ov::element::u8,      ov::element::u16,  ov::element::u32, ov::element::u64,
+    ov::element::boolean,
+    ov::element::bf16,
+    ov::element::f16,
+    ov::element::f32,
+    ov::element::f64,
+    ov::element::i4,
+    ov::element::i8,
+    ov::element::i16,
+    ov::element::i32,
+    ov::element::i64,
+    ov::element::u1,
+    ov::element::u4,
+    ov::element::u8,
+    ov::element::u16,
+    ov::element::u32,
+    ov::element::u64,
 };
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests, OVInferRequestIOTensorSetPrecisionTest,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_NPU),
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests,
+                         OVInferRequestIOTensorSetPrecisionTest,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestIOTensorSetPrecisionTest>);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests, OVInferRequestIOTensorSetPrecisionTest,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_MULTI),
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests,
+                         OVInferRequestIOTensorSetPrecisionTest,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestIOTensorSetPrecisionTest>);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests, OVInferRequestIOTensorSetPrecisionTest,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_AUTO),
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests,
+                         OVInferRequestIOTensorSetPrecisionTest,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_AUTO),
                                             ::testing::ValuesIn(autoConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestIOTensorSetPrecisionTest>);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests, OVInferRequestIOTensorSetPrecisionTestNPU,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_NPU),
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests,
+                         OVInferRequestIOTensorSetPrecisionTestNPU,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestIOTensorSetPrecisionTestNPU>);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Mutli_BehaviorTests, OVInferRequestIOTensorSetPrecisionTestNPU,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_MULTI),
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Mutli_BehaviorTests,
+                         OVInferRequestIOTensorSetPrecisionTestNPU,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestIOTensorSetPrecisionTestNPU>);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests, OVInferRequestIOTensorSetPrecisionTestNPU,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_AUTO),
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests,
+                         OVInferRequestIOTensorSetPrecisionTestNPU,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_AUTO),
                                             ::testing::ValuesIn(autoConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestIOTensorSetPrecisionTestNPU>);
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests, OVInferRequestCheckTensorPrecision,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_NPU),
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests,
+                         OVInferRequestCheckTensorPrecision,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestCheckTensorPrecision>);
 
-INSTANTIATE_TEST_SUITE_P(smoke_Multi_BehaviorTests, OVInferRequestCheckTensorPrecision,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_MULTI),
+INSTANTIATE_TEST_SUITE_P(smoke_Multi_BehaviorTests,
+                         OVInferRequestCheckTensorPrecision,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestCheckTensorPrecision>);
 
-INSTANTIATE_TEST_SUITE_P(smoke_Auto_BehaviorTests, OVInferRequestCheckTensorPrecision,
-                         ::testing::Combine(::testing::ValuesIn(prcs), ::testing::Values(ov::test::utils::DEVICE_AUTO),
+INSTANTIATE_TEST_SUITE_P(smoke_Auto_BehaviorTests,
+                         OVInferRequestCheckTensorPrecision,
+                         ::testing::Combine(::testing::ValuesIn(prcs),
+                                            ::testing::Values(ov::test::utils::DEVICE_AUTO),
                                             ::testing::ValuesIn(autoConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<OVInferRequestCheckTensorPrecision>);
