+++ src/plugins/intel_npu/tests/functional/shared_tests_instances/behavior/ov_infer_request/multithreading.cpp
@@ -2,12 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
+#include "behavior/ov_infer_request/multithreading.hpp"
+
 #include <vector>
 
-#include "behavior/ov_infer_request/multithreading.hpp"
-#include "common/utils.hpp"
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "common/utils.hpp"
+#include "intel_npu/al/config/options.hpp"
 #include "overload/ov_infer_request/multithreading.hpp"
 
 using namespace ov::test::behavior;
@@ -16,39 +17,43 @@ namespace {
 const std::vector<ov::AnyMap> configs = {{}};
 
 const std::vector<ov::AnyMap> multiConfigs = {
-        {ov::device::priorities(ov::test::utils::DEVICE_NPU),
-        ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
+    {ov::device::priorities(ov::test::utils::DEVICE_NPU), ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
 
 const std::vector<ov::AnyMap> autoConfigs = {
-        {ov::device::priorities(ov::test::utils::DEVICE_NPU),
-        ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
+    {ov::device::priorities(ov::test::utils::DEVICE_NPU), ov::device::properties(ov::test::utils::DEVICE_NPU, {})}};
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests, OVInferRequestMultithreadingTests,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests,
+                         OVInferRequestMultithreadingTests,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests, OVInferRequestMultithreadingTests,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests,
+                         OVInferRequestMultithreadingTests,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests, OVInferRequestMultithreadingTests,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests,
+                         OVInferRequestMultithreadingTests,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_AUTO),
                                             ::testing::ValuesIn(autoConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests, OVInferRequestMultithreadingTestsNPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests,
+                         OVInferRequestMultithreadingTestsNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(configs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests, OVInferRequestMultithreadingTestsNPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Multi_BehaviorTests,
+                         OVInferRequestMultithreadingTestsNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_MULTI),
                                             ::testing::ValuesIn(multiConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests, OVInferRequestMultithreadingTestsNPU,
+INSTANTIATE_TEST_SUITE_P(compatibility_smoke_Auto_BehaviorTests,
+                         OVInferRequestMultithreadingTestsNPU,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_AUTO),
                                             ::testing::ValuesIn(autoConfigs)),
                          InferRequestParamsAnyMapTestName::getTestCaseName);
