+++ src/plugins/intel_npu/tests/functional/shared_tests_instances/behavior/ov_plugin/caching_tests.cpp
@@ -3,22 +3,24 @@
 //
 
 #include "behavior/ov_plugin/caching_tests.hpp"
+
 #include <utility>
-#include "common/utils.hpp"
+
 #include "common/npu_test_env_cfg.hpp"
-#include "intel_npu/al/config/common.hpp"
+#include "common/utils.hpp"
+#include "intel_npu/al/config/options.hpp"
 
 using namespace ov::test::behavior;
 
 namespace {
 static const std::vector<ov::element::Type> nightly_precisionsNPU = {
-        ov::element::f32,
-        // ov::element::f16,
-        // ov::element::u8,
+    ov::element::f32,
+    // ov::element::f16,
+    // ov::element::u8,
 };
 
 static const std::vector<ov::element::Type> smoke_precisionsNPU = {
-        ov::element::f32,
+    ov::element::f32,
 };
 
 static const std::vector<std::size_t> batchSizesNPU = {1};
@@ -35,9 +37,12 @@ static std::vector<ovModelWithName> NPU_functions() {
     auto funcs = CompileModelCacheTestBase::getStandardFunctions();
 
     std::vector<ovModelWithName>::iterator it = remove_if(funcs.begin(), funcs.end(), [](ovModelWithName func) {
-        std::vector<std::string> bad_layers{"ReadConcatSplitAssign", "SimpleFunctionRelu",
-                                            "2InputSubtract",        "MatMulBias",
-                                            "TIwithLSTMcell1",       "KSOFunction"};
+        std::vector<std::string> bad_layers{"ReadConcatSplitAssign",
+                                            "SimpleFunctionRelu",
+                                            "2InputSubtract",
+                                            "MatMulBias",
+                                            "TIwithLSTMcell1",
+                                            "KSOFunction"};
         return std::find(bad_layers.begin(), bad_layers.end(), std::get<1>(func)) != bad_layers.end();
     });
 
@@ -46,7 +51,8 @@ static std::vector<ovModelWithName> NPU_functions() {
     return funcs;
 }
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU, CompileModelCacheTestBase,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU,
+                         CompileModelCacheTestBase,
                          ::testing::Combine(::testing::ValuesIn(smoke_functions()),
                                             ::testing::ValuesIn(smoke_precisionsNPU),
                                             ::testing::ValuesIn(batchSizesNPU),
@@ -54,7 +60,8 @@ INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU, CompileMode
                                             ::testing::Values(ov::AnyMap{})),
                          ov::test::utils::appendPlatformTypeTestName<CompileModelCacheTestBase>);
 
-INSTANTIATE_TEST_SUITE_P(nightly_BehaviorTests_CachingSupportCase_NPU, CompileModelCacheTestBase,
+INSTANTIATE_TEST_SUITE_P(nightly_BehaviorTests_CachingSupportCase_NPU,
+                         CompileModelCacheTestBase,
                          ::testing::Combine(::testing::ValuesIn(NPU_functions()),
                                             ::testing::ValuesIn(nightly_precisionsNPU),
                                             ::testing::ValuesIn(batchSizesNPU),
@@ -70,15 +77,15 @@ static std::string getTestCaseName(testing::TestParamInfo<compileModelLoadFromFi
 }
 
 const std::vector<ov::AnyMap> LoadFromFileConfigs = {
-        {ov::device::properties(ov::test::utils::DEVICE_NPU, {}),
-         ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT)},
-        {ov::device::properties(ov::test::utils::DEVICE_NPU, {}),
-         ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY)}};
+    {ov::device::properties(ov::test::utils::DEVICE_NPU, {}),
+     ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT)},
+    {ov::device::properties(ov::test::utils::DEVICE_NPU, {}),
+     ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY)}};
 
 const std::vector<std::string> TestTargets = {
-        ov::test::utils::DEVICE_AUTO,
-        ov::test::utils::DEVICE_MULTI,
-        ov::test::utils::DEVICE_BATCH,
+    ov::test::utils::DEVICE_AUTO,
+    ov::test::utils::DEVICE_MULTI,
+    ov::test::utils::DEVICE_BATCH,
 };
 
 INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_Auto_BehaviorTests_CachingSupportCase_NPU,
@@ -87,39 +94,39 @@ INSTANTIATE_TEST_SUITE_P(DISABLED_smoke_Auto_BehaviorTests_CachingSupportCase_NP
                          getTestCaseName);
 
 const std::vector<ov::AnyMap> NPULoadFromFileConfigs = {
-        {ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT)},
-        {ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY)}
-};
+    {ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT)},
+    {ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY)}};
 
 const std::vector<std::pair<ov::AnyMap, std::string>> NPUCompiledKernelsCacheTest = {
-        std::make_pair<ov::AnyMap, std::string>(
-                {ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT),
-                 ov::intel_npu::use_elf_compiler_backend(ov::intel_npu::ElfCompilerBackend::NO)},
-                "blob"),
-        std::make_pair<ov::AnyMap, std::string>(
-                {ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY),
-                 ov::intel_npu::use_elf_compiler_backend(ov::intel_npu::ElfCompilerBackend::NO)},
-                "blob"),
-        std::make_pair<ov::AnyMap, std::string>(
-                {ov::intel_npu::use_elf_compiler_backend(ov::intel_npu::ElfCompilerBackend::NO)},
-                "blob"),
-        std::make_pair<ov::AnyMap, std::string>({ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT)},
-                                                "blob"),
-        std::make_pair<ov::AnyMap, std::string>({ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY)},
-                                                "blob")
-};
+    std::make_pair<ov::AnyMap, std::string>(
+        {ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT),
+         ov::intel_npu::use_elf_compiler_backend(ov::intel_npu::ElfCompilerBackend::NO)},
+        "blob"),
+    std::make_pair<ov::AnyMap, std::string>(
+        {ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY),
+         ov::intel_npu::use_elf_compiler_backend(ov::intel_npu::ElfCompilerBackend::NO)},
+        "blob"),
+    std::make_pair<ov::AnyMap, std::string>(
+        {ov::intel_npu::use_elf_compiler_backend(ov::intel_npu::ElfCompilerBackend::NO)},
+        "blob"),
+    std::make_pair<ov::AnyMap, std::string>({ov::hint::performance_mode(ov::hint::PerformanceMode::THROUGHPUT)},
+                                            "blob"),
+    std::make_pair<ov::AnyMap, std::string>({ov::hint::performance_mode(ov::hint::PerformanceMode::LATENCY)}, "blob")};
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU, CompileModelLoadFromFileTestBase,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU,
+                         CompileModelLoadFromFileTestBase,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(NPULoadFromFileConfigs)),
                          getTestCaseName);
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU, CompileModelLoadFromMemoryTestBase,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU,
+                         CompileModelLoadFromMemoryTestBase,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(NPULoadFromFileConfigs)),
                          ov::test::utils::appendPlatformTypeTestName<CompileModelLoadFromMemoryTestBase>);
 
-INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU, CompiledKernelsCacheTest,
+INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CachingSupportCase_NPU,
+                         CompiledKernelsCacheTest,
                          ::testing::Combine(::testing::Values(ov::test::utils::DEVICE_NPU),
                                             ::testing::ValuesIn(NPUCompiledKernelsCacheTest)),
                          ov::test::utils::appendPlatformTypeTestName<CompiledKernelsCacheTest>);
